requirements.txt
aiogram==3.3.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.9
aiohttp==3.9.3
websockets==12.0
pydantic==2.6.1
pydantic-settings==2.1.0 